Creative mode preventing in Multicrates

+++ src/main/java/me/ztowne13/customcrates/crates/options/CMultiCrateInventory.java
@@ -17,6 +17,7 @@ import me.ztowne13.customcrates.utils.ChatUtils;
 import me.ztowne13.customcrates.utils.CrateUtils;
 import me.ztowne13.customcrates.utils.Utils;
 import org.bukkit.Bukkit;
+import org.bukkit.GameMode;
 import org.bukkit.Material;
 import org.bukkit.configuration.file.FileConfiguration;
 import org.bukkit.entity.Player;
@@ -383,32 +384,38 @@ public class CMultiCrateInventory extends CSetting
 			}
 			else if(pm.isUseVirtualCrate() ? pm.getPdm().getVCCrateData(clickedCrate).getCrates() > 0 : true)
 			{
-
-				if (crate.getCs().getCmci().getCrateSpots().get(slot).getCs().getCh().tick(p, pm.getLastOpenCrate(), CrateState.OPEN, false))
+				if(!p.getGameMode().equals(GameMode.CREATIVE) || (Boolean) cc.getSettings().getConfigValues().get("open-creative"))
 				{
-					if(pm.isUseVirtualCrate())
+					if (crate.getCs().getCmci().getCrateSpots().get(slot).getCs().getCh().tick(p, pm.getLastOpenCrate(), CrateState.OPEN, false))
 					{
-						pm.getPdm().setVirtualCrateCrates(clickedCrate, pm.getPdm().getVCCrateData(clickedCrate).getCrates() - 1);
-					}
-					else if (!crates.getCs().getOt().equals(ObtainType.STATIC))
-					{
-						try
+						if(pm.isUseVirtualCrate())
 						{
-							PlacedCrate pc = PlacedCrate.get(cc, pm.getLastOpenCrate());
-							pc.delete();
+							pm.getPdm().setVirtualCrateCrates(clickedCrate, pm.getPdm().getVCCrateData(clickedCrate).getCrates() - 1);
 						}
-						catch(Exception exc)
+						else if (!crates.getCs().getOt().equals(ObtainType.STATIC))
 						{
+							try
+							{
+								PlacedCrate pc = PlacedCrate.get(cc, pm.getLastOpenCrate());
+								pc.delete();
+							}
+							catch(Exception exc)
+							{
 
+							}
+						}
+					}
+					else
+					{
+						if (p.getLocation().distance(pm.getLastOpenCrate()) > 10)
+						{
+							p.closeInventory();
 						}
 					}
 				}
 				else
 				{
-					if (p.getLocation().distance(pm.getLastOpenCrate()) > 10)
-					{
-						p.closeInventory();
-					}
+					Messages.DENY_CREATIVE_MODE.msgSpecified(cc, p);
 				}
 			}
 			else
